'Fix' problem where base/cutting layer isn't available.
This can happen when the client selects a subset of layers which don't include both the base and cutting layers. If the base layer isn't selected, then the appropriate thing to do, which this patch does, is skip the post-processing step.

However, if the client selects the base layer, but not the cutting layer, then it's too late to do anything about it in the post-processor. The appropriate thing may be to not perform the cutting process, or it might be to change the code so that it additionally selects the cutting layer and filters it out when the client doesn't want it.