Fix a bug where we were treating an int as a str.
The `to_float` function expects a `str` or None, but we were
passing it an `int`. The function then tries to replace ';' and
',', but that's not a valid operation on an `int`. Instead, where
the parameter is missing, we'll default to None which is handled
by `to_float` and apply the `int` default afterwards.