Allow overlap post-processor to keep features in the same layer.
This is signaled by setting `new_layer_name` to None. Also added
a more flexible way to build the dictionary of properties on the
new feature.

Finally, added a feature to buffer geometries - although it turns
out this isn't useful for water polygons it might be useful for
other layers in the future.