Better geometry type comparison.
Previously, the comparison had been on Python type. This seemed to
work most of the time, but there were some odd instances of things
claiming to be of `MultiLineString` type, but where
`isinstance(obj, BaseMultipartGeometry)` was `False`. I have a
feeling it might be something `ctypes` related.

It turns out that Shapely doesn't expose the geometry type ID, so
we still have to do a stringly comparison, but I've tried to make
it as simple as possible and do it only once.

This is the base cause of mapzen/vector-datasource#204: Objects
were returned from the intercut algorithm which were
MultiLineStrings, but which were not being detected as such in the
`isinstance` test. This led to some road segments being dropped
from the tile.