Added post-process filter to remove duplicate features.
Where being a duplicate is defined by having some subset of
properties in common and being within a certain distance of the
"original" feature.

Feature order is not changed, so more features should be sorted
into importance order by the sorting step which runs at the end of
the filter functions.

The type of feature to filter out may be selected using the
geometry filter. This is primarily of use when filtering out
duplicate POIs / labels only.